% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatentSafe.R
\name{create_metadata_xml}
\alias{create_metadata_xml}
\title{Create a Metadata XML packet}
\usage{
create_metadata_xml(named_list)
}
\arguments{
\item{named_list}{key value pairs of metadata items}
}
\value{
the metadata XML for PatentSafe submission
}
\description{
Create a Metadata XML packet
}
