% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatentSafe.R
\name{create_valid_url}
\alias{create_valid_url}
\title{Clean up the Server URL}
\usage{
create_valid_url(url)
}
\arguments{
\item{A}{character string of the server URL or domain name}
}
\value{
A string with the user URL in https:// format
}
\description{
This function takes a variety of URL and domain name formats and
creates an https:// version
}
\keyword{internal}
