% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatentSafe.R
\name{open_patentsafe_document}
\alias{open_patentsafe_document}
\title{Open a PatentSafe document in the browser, based on the return code from
PatentSafe's HTTP endpoint}
\usage{
open_patentsafe_document(submission_return, base_url)
}
\arguments{
\item{submission_return}{The return from a PatentSafe submission HTTP call}

\item{base_url}{The PatentSafe server's URL}
}
\description{
Open a PatentSafe document in the browser, based on the return code from
PatentSafe's HTTP endpoint
}
\keyword{internal}
