% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatentSafe.R
\name{openPatentSafeDocument}
\alias{openPatentSafeDocument}
\title{Title Open a PatentSafe document in the browser, based on the}
\usage{
openPatentSafeDocument(submissionReturn, baseURL)
}
\arguments{
\item{submissionReturn}{The return from a PatentSafe submission HTTP call}

\item{baseURL}{The PatentSafe server's URL}
}
\value{
This function isn't exported so there's no @export
}
\description{
Title Open a PatentSafe document in the browser, based on the
}
\examples{
openPatentSafeDocument('report.pdf', 'https://demo.morescience.com', 'clarusc')
}
