% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatentSafe.R
\name{submit}
\alias{submit}
\title{Title}
\usage{
submit(url, authorId, reportFilename)
}
\arguments{
\item{url}{the URL of your PatentSafe Server}

\item{authorId}{your PatentSafe user ID}

\item{reportFilename}{filename of your report file (should be a PDF)}
}
\value{
PatentSafe document ID or error code
}
\description{
Title
}
\examples{
submit('https://demo.morescience.com', 'clarusc', 'report.pdf')
}
