% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatentSafe.R
\name{submit_pdf}
\alias{submit_pdf}
\title{Submit a PDF}
\usage{
submit_pdf(
  report_filename,
  text_content = NULL,
  summary = NULL,
  metadata = NULL,
  destination = "sign",
  attachment_filename = NULL,
  url = Sys.getenv("PATENTSAFE_URL"),
  author_id = Sys.getenv("PATENTSAFE_USERID")
)
}
\arguments{
\item{report_filename}{filename of your report file (should be a PDF)}

\item{text_content}{Textual content of the PDF, defaults to PatentSafe's
automatic extraction}

\item{summary}{Summary of the experiment, defaults to PatentSafe's automatic
calculation}

\item{metadata}{Metadata for the PatentSafe document, as a}

\item{destination}{PatentSafe queue to submit to, defaults to "sign"}

\item{attachment_filename}{The filename of an attachment defaults to
no attachment}

\item{url}{the URL of your PatentSafe Server defaults to PATENTSAFE_USERID}

\item{author_id}{your PatentSafe user ID defaults to PATENTSAFE_URL}
}
\value{
PatentSafe document ID or error code
}
\description{
Submit a PDF
}
\examples{
# Path to the PDF 
\dontrun{
submit_pdf(test.pdf,
           url = "https://demo.morescience.com",
           author_id = "clarusc")
}
}
