% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatentSafe.R
\name{submit_rmd}
\alias{submit_rmd}
\title{Submit an .Rmd file to PatentSafe. This is the best option for
submission because we can extract text etc.}
\usage{
submit_rmd(
  report_filename = "Writeup.Rmd",
  summary = NULL,
  metadata = NULL,
  destination = "sign",
  attachment_filename = NULL,
  url = Sys.getenv("PATENTSAFE_URL"),
  author_id = Sys.getenv("PATENTSAFE_USERID")
)
}
\arguments{
\item{report_filename}{Filename of the Report file, defaults to Report.Rmd}

\item{summary}{Summary of the experiment, defaults to PatentSafe's
automatic calculation}

\item{metadata}{Metadata for the PatentSafe document, as a}

\item{destination}{PatentSafe queue to submit to, defaults to "sign"}

\item{attachment_filename}{The filename of an attachment defaults to
no attachment}

\item{url}{the URL of your PatentSafe Server defaults to PATENTSAFE_USERID}

\item{author_id}{your PatentSafe user ID defaults to PATENTSAFE_URL}
}
\value{
Error code or opens PatentSafe in a browser
}
\description{
Submit an .Rmd file to PatentSafe. This is the best option for
submission because we can extract text etc.
}
\examples{
\dontrun{
submit_rmd('Report.Rmd',
           'https://demo.morescience.com',
           'clarusc')
}
}
