% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatentSafe.R
\name{submit_this_project}
\alias{submit_this_project}
\title{Submit this project to PatentSafe as both a PDF and a .zip attachment}
\usage{
submit_this_project(
  directory = ".",
  report_filename = "Writeup.Rmd",
  summary = NULL,
  metadata = NULL,
  destination = "sign",
  url = Sys.getenv("PATENTSAFE_URL"),
  author_id = Sys.getenv("PATENTSAFE_USERID")
)
}
\arguments{
\item{directory}{The directory to submit. Must contain a .Rmd file}

\item{report_filename}{Filename of the Report file, defaults to Report.Rmd}

\item{summary}{Summary of the experiment, defaults to PatentSafe's automatic
calculation}

\item{metadata}{Metadata for the PatentSafe document, as a}

\item{destination}{PatentSafe queue to submit to, defaults to "sign"}

\item{url}{the URL of your PatentSafe Server defaults to PATENTSAFE_USERID}

\item{author_id}{your PatentSafe user ID defaults to PATENTSAFE_URL}
}
\value{
Error code or opens PatentSafe in a browser
}
\description{
Submit this project to PatentSafe as both a PDF and a .zip attachment
}
\examples{
submit_this_project(system.file("extdata", package = 'PatentSafeR'),
                    url = 'https://demo.morescience.com',
                    author_id = "clarusc")
}
