% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatentSafe.R
\name{submitPDF}
\alias{submitPDF}
\title{Submit a PDF}
\usage{
submitPDF(
  reportFilename,
  textContent = NULL,
  summary = NULL,
  metadata = NULL,
  destination = "sign",
  attachmentFilename = NULL,
  url = Sys.getenv("PATENTSAFE_URL"),
  authorId = Sys.getenv("PATENTSAFE_USERID")
)
}
\arguments{
\item{reportFilename}{filename of your report file (should be a PDF)}

\item{textContent}{Textual content of the PDF, defaults to PatentSafe's automatic extraction}

\item{summary}{Summary of the experiment, defaults to PatentSafe's automatic calculation}

\item{metadata}{Metadata for the PatentSafe document, as a}

\item{destination}{PatentSafe queue to submit to, defaults to "sign"}

\item{attachmentFilename}{The filename of an attachment defaults to no attachment}

\item{url}{the URL of your PatentSafe Server defaults to PATENTSAFE_USERID}

\item{authorId}{your PatentSafe user ID defaults to PATENTSAFE_URL}
}
\value{
PatentSafe document ID or error code
}
\description{
Submit a PDF
}
\examples{
submitPDF('report.pdf', 'https://demo.morescience.com', 'clarusc')
}
