% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatentSafe.R
\name{submitRmd}
\alias{submitRmd}
\title{Submit an .Rmd file to PatentSafe. This is the best option for submission because we can extract text etc.}
\usage{
submitRmd(
  reportFilename,
  summary = NULL,
  metadata = NULL,
  destination = "sign",
  url = Sys.getenv("PATENTSAFE_URL"),
  authorId = Sys.getenv("PATENTSAFE_USERID")
)
}
\arguments{
\item{reportFilename}{Filename of the Report file, defaults to Report.Rmd}

\item{summary}{Summary of the experiment, defaults to PatentSafe's automatic calculation}

\item{metadata}{Metadata for the PatentSafe document, as a}

\item{destination}{PatentSafe queue to submit to, defaults to "sign"}

\item{url}{the URL of your PatentSafe Server defaults to PATENTSAFE_USERID}

\item{authorId}{your PatentSafe user ID defaults to PATENTSAFE_URL}
}
\value{
Error code or opens PatentSafe in a browser
}
\description{
Submit an .Rmd file to PatentSafe. This is the best option for submission because we can extract text etc.
}
\examples{
submitPDF('report.pdf', 'https://demo.morescience.com', 'clarusc')
}
