% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatentSafe.R
\name{create_metadata_xml}
\alias{create_metadata_xml}
\title{Create a PatentSafe Metadata XML packet}
\usage{
create_metadata_xml(named_list)
}
\value{
A string with the metadata XML for PatentSafe submission
}
\description{
@param named_list A list of key value pairs of metadata items
}
\keyword{internal}
